requirements.txt
numpy==1.20.3
matplotlib==3.3.4
torchvision==0.10.0
scikit_image==0.18.1
albumentations==1.2.1
scipy==1.6.2
opencv_python_headless==4.6.0.66
pandas==1.3.0
torch==1.9.0
tqdm==4.61.2
torchsummary==1.5.1
Pillow==9.2.0
scikit_learn==1.1.3
skimage==0.0
